feat: allow node paths to be overridden
This adds the `node_runtime` settings, which allows the user to override
the location of `node`, `npm` and the npm cache. This is primarily for
supporting NixOS.

This changes behavior slightly: the blank npmrc files are stored in the
cache directory, not the node bin directory (as that is immutable on
NixOS).

A sync+async mutex crate is not currently a dependency in the workspace,
so a mpsc is used to smuggle new configuration into the singleton -
using `futures::executor::block_on` on the mutex seems to deadlock the
GUI thread.